Move some vector functions around a bit with tests for vectorutils.
