+++ doc/python/source/rsgislib_vectorattrs.rst
@@ -23,6 +23,7 @@ Add Columns
 Calculate Column Values
 --------------------------
 .. autofunction:: rsgislib.vectorattrs.pop_bbox_cols
+.. autofunction:: rsgislib.vectorattrs.add_geom_bbox_cols
 .. autofunction:: rsgislib.vectorattrs.add_unq_numeric_col
 
 Get Column Summaries
