+++ python_tests/test_classification.py
@@ -19,6 +19,12 @@ try:
 except ImportError:
     KALEIDO_NOT_AVAIL = True
 
+GEOPANDAS_NOT_AVAIL = False
+try:
+    import geopandas
+except ImportError:
+    GEOPANDAS_NOT_AVAIL = True
+
 DATA_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data")
 ZONALSTATS_DATA_DIR = os.path.join(DATA_DIR, "zonalstats")
 CLASSIFICATION_DATA_DIR = os.path.join(DATA_DIR, "classification")
@@ -363,3 +369,19 @@ def test_plot_train_data(tmp_path):
     plot_files = glob.glob(os.path.join(tmp_path, "*.png"))
 
     assert len(plot_files) == 100
+
+
+@pytest.mark.skipif(GEOPANDAS_NOT_AVAIL, reason="geopandas dependency not available")
+def test_create_acc_pt_sets(tmp_path):
+    import rsgislib.classification
+    import glob
+    vec_file = os.path.join(CLASSIFICATION_DATA_DIR, "gmw_acc_roi_30_cls_acc_pts.geojson")
+    vec_lyr = "gmw_acc_roi_30_cls_acc_pts"
+
+    out_vec_file_base = os.path.join(tmp_path, "gmw_acc_roi_30_cls_acc_pts_")
+
+    rsgislib.classification.create_acc_pt_sets(vec_file, vec_lyr, out_vec_file_base, "gmw_acc_roi_30_cls_acc_pts", "gmw_v25_cls", 10, sets_col="set_id", out_format="GeoJSON", out_ext="geojson", shuffle_rows=True, rnd_seed=42)
+    assert len(glob.glob(os.path.join(tmp_path, "gmw_acc_roi_30_cls_acc_pts_*.geojson"))) == 10
+
+
+
